FactoryGirl.lint can lint traits
Why?

Traits are another way to extend or modify behavior of a factory;
because of this, applied traits may result in invalid factories, even if
the base factory is valid.

To enable this functionality, use:

    FactoryGirl.lint traits:true

Note that, because FactoryGirl uses the `create` strategy when linting,
this may greatly slow the time to lint factories.