Update bundler before in Travis CI builds
Builds are broken because the `bundler` version that Travis CI
automatically uses at this time (`1.7.6`) is coming across an
error that was fixed in a later version/release.

Having an outdated version of `bundler` in the Travis CI builds is an
ongoing issue and this is currently the easiest work-around.

See:

* https://github.com/bundler/bundler/issues/3558
* https://github.com/travis-ci/travis-ci/issues/3531

[fixes #846]