Name first arg of DefinitionProxy#method_missing
This code was a bit confusing before this change.

The only time we care about `args.first` is if we are creating an
association, so this PR gives `args.first` the name
`association_options` to reflect that.

The PR also pulls out a predicate method to give a name to the step of
validating that `args.first` are in indeed association options.