Update arity checks for Symbol#to_proc change
Ruby has [changed the arity of Symbol#to_proc from -1 to -2][ruby]. This commit
updates some arity checks we have in factory_bot to handle that case.
This should get the Ruby HEAD + Rails 6 tests passing on CI (the Ruby
head with other versions of Rails is still failing for various reasons).

[ruby]: https://github.com/ruby/ruby/commit/5b29ea0845c14092abd866ce0183c52635bade4c